Fix warnings

+++ src/NetBox/FarDialog.cpp
@@ -845,7 +845,7 @@ LONG_PTR TFarDialog::SendMessage(int Msg, intptr_t Param1, LONG_PTR Param2)
     Msg, static_cast<int>(Param1), Param2);
 }
 //---------------------------------------------------------------------------
-uintptr_t TFarDialog::GetSystemColor(int Index)
+uintptr_t TFarDialog::GetSystemColor(intptr_t Index)
 {
   return static_cast<uintptr_t>(GetFarPlugin()->FarAdvControl(ACTL_GETCOLOR, reinterpret_cast<void *>(Index)));
 }
